requirements.txt
discord.py
aiohttp
psutil